pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-index"
version = "0.1.0"
description = "Hierarchical tree index for agentic document retrieval"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "httpx>=0.27",
    "google-genai>=1.0",
    "click>=8.1",
    "pydantic>=2.0",
    "beautifulsoup4>=4.12",
    "markdownify>=0.14",
    "mcp[cli]>=1.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "rank-bm25>=0.2.2",
]

[project.optional-dependencies]
eval = [
    "chromadb>=0.5",
    "sentence-transformers>=3.0",
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "scipy>=1.14",
    "numpy>=1.26",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
agentic-index = "agentic_index.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_index"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
